requirements.txt
google-generativeai==0.4.0
pillow==10.2.0
streamlit==1.31.0
streamlit-option-menu==0.3.12
protobuf~=4.25.6